serialize team dto for project update

+++ server/api/projects/teams.py
@@ -6,9 +6,28 @@ from server.services.users.authentication_service import token_auth, tm
 
 
 class ProjectsTeamsAPI(Resource):
-    @tm.pm_only(False)
     def get(self, project_id):
-        """ Get projects associated with the project"""
+        """ Get teams assigned with a project
+        ---
+        tags:
+          - teams
+        produces:
+          - application/json
+        parameters:
+            - name: project_id
+              in: path
+              description: Unique project ID
+              required: true
+              type: integer
+              default: 1
+        responses:
+            200:
+                description: Teams listed successfully
+            404:
+                description: Not found
+            500:
+                description: Internal Server Error
+        """
         try:
             teams_dto = TeamService.get_project_teams_as_dto(project_id)
             return teams_dto.to_primitive(), 200
@@ -19,38 +38,141 @@ class ProjectsTeamsAPI(Resource):
 
     @tm.pm_only(False)
     @token_auth.login_required
-    def put(self, team_id, project_id):
-        """ Assign team to the project"""
+    def post(self, team_id, project_id):
+        """ Assign a team to a project
+        ---
+        tags:
+          - teams
+        produces:
+          - application/json
+        parameters:
+            - in: header
+              name: Authorization
+              description: Base64 encoded session token
+              required: true
+              type: string
+              default: Token sessionTokenHere==
+            - name: project_id
+              in: path
+              description: Unique project ID
+              required: true
+              type: integer
+              default: 1
+            - name: team_id
+              in: path
+              description: Unique team ID
+              required: true
+              type: integer
+              default: 1
+            - in: body
+              name: body
+              required: true
+              description: The role that the team will have on the project
+              schema:
+                  properties:
+                      role:
+                        type: string
+        responses:
+            201:
+                description: Team project assignment created
+            401:
+                description: Forbidden, if user is not a manager of the project
+            403:
+                description: Forbidden, if user is not authenticated
+            404:
+                description: Not found
+            500:
+                description: Internal Server Error
+        """
+        if not TeamService.user_is_manager(team_id, tm.authenticated_user_id):
+            return {"Error": "User is not an admin or a manager for the team"}, 401
+
         try:
-            TeamService.add_team_project(team_id, project_id)
-            return {"Success": "Project added"}, 200
+            role = request.get_json(force=True)["role"]
+        except DataError as e:
+            current_app.logger.error(f"Error validating request: {str(e)}")
+            return str(e), 400
+
+        try:
+            TeamService.add_team_project(team_id, project_id, role)
+            return (
+                {
+                    "Success": "Team {} assigned to project {} with role {}".format(
+                        team_id, project_id, role
+                    )
+                },
+                201,
+            )
         except Exception as e:
-            error_msg = f"Team PUT - unhandled error: {str(e)}"
+            error_msg = f"Project Team POST - unhandled error: {str(e)}"
             current_app.logger.critical(error_msg)
             return {"Error": error_msg}, 500
 
     @tm.pm_only(False)
     @token_auth.login_required
     def patch(self, team_id, project_id):
-        """ Update role of the project-team"""
+        """ Update role of a team on a project
+        ---
+        tags:
+          - teams
+        produces:
+          - application/json
+        parameters:
+            - in: header
+              name: Authorization
+              description: Base64 encoded session token
+              required: true
+              type: string
+              default: Token sessionTokenHere==
+            - name: project_id
+              in: path
+              description: Unique project ID
+              required: true
+              type: integer
+              default: 1
+            - name: team_id
+              in: path
+              description: Unique team ID
+              required: true
+              type: integer
+              default: 1
+            - in: body
+              name: body
+              required: true
+              description: The role that the team will have on the project
+              schema:
+                  properties:
+                      role:
+                        type: string
+        responses:
+            201:
+                description: Team project assignment created
+            401:
+                description: Forbidden, if user is not a manager of the project
+            403:
+                description: Forbidden, if user is not authenticated
+            404:
+                description: Not found
+            500:
+                description: Internal Server Error
+        """
+        if not TeamService.user_is_manager(team_id, tm.authenticated_user_id):
+            return {"Error": "User is not an admin or a manager for the team"}, 401
         try:
             role = request.get_json(force=True)["role"]
-
-            if not TeamService.user_is_manager(team_id, tm.authenticated_user_id):
-                return {"Error": "User is not a admin or a manager for the team"}, 401
         except DataError as e:
-            current_app.logger.error(f"error validating request: {str(e)}")
+            current_app.logger.error(f"Error validating request: {str(e)}")
             return str(e), 400
 
         try:
             TeamService.change_team_role(team_id, project_id, role)
-            return {"Status": "Updated"}, 200
+            return {"Status": "Team role updated successfully."}, 200
         except NotFound as e:
             return {"Error": str(e)}, 404
         except TeamServiceError as e:
             return str(e), 402
         except Exception as e:
-            error_msg = f"Team PATCH - unhandled error: {str(e)}"
+            error_msg = f"Team-Project PATCH - unhandled error: {str(e)}"
             current_app.logger.critical(error_msg)
             return {"Error": error_msg}, 500
 
@@ -58,7 +180,7 @@ class ProjectsTeamsAPI(Resource):
     @token_auth.login_required
     def delete(self, team_id, project_id):
         """
-        Deletes the specified team project
+        Deletes the specified team project assignment
         ---
         tags:
           - teams
@@ -79,16 +201,19 @@ class ProjectsTeamsAPI(Resource):
               default: 1
         responses:
             200:
-                description: Team project deleted
+                description: Team unassigned of the project
+            401:
+                description: Forbidden, if user is not a manager of the project
             403:
-                description: Forbidden, if user attempting to ready other messages
+                description: Forbidden, if user is not authenticated
             404:
                 description: Not found
             500:
                 description: Internal Server Error
         """
+        if not TeamService.user_is_manager(team_id, tm.authenticated_user_id):
+            return {"Error": "User is not an admin or a manager for the team"}, 401
         try:
-
             TeamService.delete_team_project(team_id, project_id)
             return {"Success": True}, 200
         except NotFound:
